Move component to components folder
