+++ src/App.jsx
@@ -1,17 +1,16 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-
-import Header from './Header.jsx';
+import React from 'react'
+import ReactDOM from 'react-dom'
+import Header from './components/Header/index.jsx'
 
 const App = () => {
   return (
     <div className='container'>
-      <Header/>
-      <div className="container-view">Some activities should be here</div>
+      <Header />
+      <div className='container-view'>Some activities should be here</div>
     </div>
-  );
-};
+  )
+}
 
-ReactDOM.render(<App/>, document.getElementById('app'));
+ReactDOM.render(<App />, document.getElementById('app'))
 
-export default App;
+export default App
